# cache chalk queries