@flodiebold 2019-06-26
LGTM. I didn't do it at first because it's supposed to be a simple transformation and I didn't want to cache everything twice, but the numbers speak for it...