@matklad 2019-06-26
One more important bit of info here is that after this change profile points squarely towards the guts of chalk:

https://salsa.zulipchat.com/#narrow/stream/184114-rfcs.2Fdrafts-and-proposals/topic/mutable.20constants/near/169015909